+++ edge/Queue.ts
@@ -3,26 +3,48 @@ import * as vscode from 'vscode'
 
 import Log from './Log'
 
-const pendingActionList: Array<{ action: (options?) => Promise<any>, options?}> = []
+type Action = (options?: object) => Promise<any>
 
-export function put(action: (options?: object) => Promise<any>) {
-	return async (options?: object) => {
+const pendingActionList: Array<{ action: Action, options: object }> = []
+
+let cancellationService: vscode.CancellationTokenSource
+
+export function put(action: Action, bypassActionList?: Array<Action>) {
+	return async (options: object = {}) => {
 		try {
+			// Do not enqueue the same operation
 			if (_.isEqual(pendingActionList[0], { action, options })) {
 				return undefined
+			}
+
+			// Cancel the active operation if the new pending operation requests cancellation
+			if (pendingActionList.length > 0 && _.includes(bypassActionList, _.last(pendingActionList).action)) {
+				if (cancellationService) {
+					cancellationService.cancel()
+					cancellationService = null
+				}
+				pendingActionList.pop()
+
+				while (pendingActionList.length > 0 && _.includes(bypassActionList, _.last(pendingActionList).action)) {
+					pendingActionList.pop()
+				}
 			}
 
 			pendingActionList.unshift({ action, options })
 
 			if (pendingActionList.length === 1) {
-				await action(options)
+				cancellationService = new vscode.CancellationTokenSource()
+				await action({ ...options, token: cancellationService.token })
 				pendingActionList.pop()
 
 				while (pendingActionList.length > 0) {
-					const { action: nextAction, options } = _.last(pendingActionList)
-					await nextAction(options)
+					cancellationService = new vscode.CancellationTokenSource()
+					const { action, options } = _.last(pendingActionList)
+					await action({ ...options, token: cancellationService.token })
 					pendingActionList.pop()
 				}
+
+				cancellationService = null
 			}
 
 		} catch (ex) {
@@ -42,4 +64,5 @@ export function run(action: (options?: object) => Promise<any>) {
 
 export function clear() {
 	pendingActionList.splice(0, pendingActionList.length)
+	cancellationService = null
 }