+++ edge/sleep.ts
@@ -1,0 +1,7 @@
+export default async function (time: number) {
+	return new Promise(resolve => {
+		setTimeout(() => {
+			resolve()
+		}, time)
+	})
+}