+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2.7.1
 - Added _--no-pager_ for all Git commands.
+- Amended `gitGrace.checkout` command so it uses the new _QuickPick_ API from _vscode_.
 
 ### 2.7.0
 - Added `gitGrace.commitSmart` command.