+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-### 2.7.1
+### 2.8.0
 - Added _--no-pager_ for all Git commands.
 - Amended `gitGrace.checkout` command so it uses the new _QuickPick_ API from _vscode_.
+- Amended `gitGrace.checkout` command so it cancels the pending `gitGrace.fetch` command.
 
 ### 2.7.0
 - Added `gitGrace.commitSmart` command.